Add daily midnight EST cron job to repopulate recurring tasks
Extracts the recurring task backfill logic from the API route into a
shared module (server/lib/recurringTaskBackfill.ts) and adds a cron job
that runs at 12:00 AM EST every day to automatically create new task
instances for all recurring series. This ensures recurring tasks
repopulate at the start of each day without manual intervention.
